finish device searching
